Split bot setup into separate module
The `TextifyBot.py` script will now be only the main driver for the program, rather than performing all tasks.
